requirement.txt
Flask==2.2.2
Werkzeug==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
requests==2.26.0
pytest==6.2.4
psycopg2-binary>=2.9.0

